Fix : Id 타입 수정에 따른 관련 클래스 수정 - DR

+++ Spring/src/main/java/com/heaildairy/www/emotion/dto/MoodDetailDTO.java
@@ -8,7 +8,7 @@ import java.math.BigDecimal;
 @Data
 @NoArgsConstructor
 public class MoodDetailDTO {
-    private Integer detailId;
+    private Long detailId;
     private Integer percentage;
     private String emotionType;
     private Long diaryId;
