+++ Spring/src/main/java/com/heaildairy/www/emotion/repository/MoodDetailRepository.java
@@ -3,5 +3,5 @@ package com.heaildairy.www.emotion.repository;
 import com.heaildairy.www.emotion.entity.MoodDetail;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MoodDetailRepository extends JpaRepository<MoodDetail, Integer> {
+public interface MoodDetailRepository extends JpaRepository<MoodDetail, Long> {
 }
